разбил на модули. Сменил декларативные функции на стрелочные
